+++ arctic_training/checkpoint/hf_engine.py
@@ -22,7 +22,7 @@ import deepspeed
 import safetensors as sf
 from deepspeed.runtime.zero.partition_parameters import ZeroParamStatus
 
-from arctic_training.checkpoint import CheckpointEngine
+from arctic_training.checkpoint.engine import CheckpointEngine
 
 # number of parameters per checkpoint shard
 SHARD_SIZE = 2e9
