+++ arctic_training/scheduler/__init__.py
@@ -12,6 +12,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from arctic_training.scheduler.factory import SchedulerFactory
-from arctic_training.scheduler.hf_factory import HFSchedulerFactory
