+++ arctic_training/trainer/__init__.py
@@ -12,6 +12,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from arctic_training.trainer.sft_trainer import SFTTrainer
-from arctic_training.trainer.trainer import Trainer
